+++ Sources/Engine/Sound/Wave.h
@@ -27,15 +27,15 @@ private:
   // Wave data
   WAVEFORMATEX pwi_wfeWave;
   WAVEFORMATEX pwi_wfeDesired;
-  ULONG  pwi_ulRiffLength, pwi_ulDataLength;
+  uint32_t  pwi_ulRiffLength, pwi_ulDataLength;
   BOOL   pwi_bInfoLoaded,  pwi_bDataLoaded; // Status
   SWORD *pwi_pswMemory; // Memory
 
   /* Conversion */
   DOUBLE pwi_dRatio;
   // get and store data
-  inline ULONG GetData_t( CTStream *pCstrInput);
-  inline void  StoreData( ULONG ulData);
+  inline uint32_t GetData_t( CTStream *pCstrInput);
+  inline void  StoreData( uint32_t ulData);
   void CopyData_t(   CTStream *pCstrInput);
   void ShrinkData_t( CTStream *pCstrInput);
 
@@ -51,14 +51,14 @@ public:
   void LoadData_t( CTStream *pCstrInput, SWORD *pswMemory, WAVEFORMATEX &SwfeDesired);
 
   /* Length in bytes / blocks / seconds */
-  ULONG  GetByteLength(void);
-  ULONG  GetDataLength(void);
-  ULONG  GetDataLength( WAVEFORMATEX SwfeDesired);
+  uint32_t  GetByteLength(void);
+  uint32_t  GetDataLength(void);
+  uint32_t  GetDataLength( WAVEFORMATEX SwfeDesired);
   DOUBLE GetSecondsLength(void);
 
   /* Buffer length in bytes */
-  ULONG DetermineBufferSize(void);
-  ULONG DetermineBufferSize( WAVEFORMATEX SwfeDesired);
+  uint32_t DetermineBufferSize(void);
+  uint32_t DetermineBufferSize( WAVEFORMATEX SwfeDesired);
 };
